requirements.txt
requests>=2.31.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
tqdm>=4.66.0
lxml>=4.9.0
Jinja2>=3.1.0
Flask>=3.0.0
sumy>=0.11.0
nltk>=3.8.0
deep-translator>=1.11.0
pdfplumber>=0.10.0